Stage/requirements.txt
alembic==0.9.2
amqp==2.1.4
apturl==0.5.2
Babel==2.4.0
bcrypt==3.1.3
beautifulsoup4==4.4.1
billiard==3.5.0.2
blinker==1.3
Brlapi==0.6.4
celery==4.0.2
cffi==1.10.0
chardet==2.3.0
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
defer==1.0.6
feedparser==5.1.3
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Celery==2.4.3
Flask-Celery-Helper==1.1.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemyPlus==0.7.5
Flask-WTF==0.14.2
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.2
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==2.0.2
mpmath==0.19
mysqlclient==1.3.10
numpy==1.12.1
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pandas==0.20.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
ptyprocess==0.5
pyasn1==0.1.9
pycparser==2.17
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.6.0
python-debian==0.1.27
python-editor==1.0.3
python-systemd==231
pytz==2017.2
pyxdg==0.25
redis==2.10.5
reportlab==3.3.0
requests==2.9.1
rpy2==2.8.5
scipy==0.19.0
sessioninstaller==0.0.0
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.10
sympy==1.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
vine==1.1.3
Werkzeug==0.12.2
Whoosh==2.7.4
WTForms==2.1
xdiagnose==3.8.4.1
xkit==0.0.0
xlrd==1.0.0
XlsxWriter==0.7.3
flask-socketio
gevent

